完善activator

+++ src/client/extend/activator.ts
@@ -4,23 +4,29 @@ import { ExtensionManager, IExtensionIdentifier, IMainExtension } from './extend
 
 type extensionId = string
 
-export interface extensionInstance {
+/**
+ * @description activate是插件激活时执行的函数,
+ * beforeClose会在结束插件结束之前执行,
+ * 而workerEntrance提供一个需要做cpu密集型工作的js模块的绝对路径
+ */
+export interface IExtensionInstance {
     activate: () => void
-    beforeClose: () => boolean
-    actualEntrance: string
+    beforeClose: () => void
+    workerEntrance: string | undefined | null
 }
 
-export interface extensionInstanceManager {
+export interface IExtensionInstanceManager {
     identifier: IExtensionIdentifier
-    worker: Worker
-    instance: extensionInstance
+    worker: Worker | undefined | null
+    instance: IExtensionInstance
 }
 
 export class ExtensionActivator {
     static events: EventEmitter
-    static extensionInstanceManagers: Map<extensionId, extensionInstanceManager>
+    static extensionInstanceManagers: Map<extensionId, IExtensionInstanceManager>
 
     constructor() {
+        ExtensionActivator.events = new EventEmitter()
         ExtensionActivator.extensionInstanceManagers = new Map()
     }
 
@@ -28,18 +34,25 @@ export class ExtensionActivator {
         ExtensionActivator.events.emit('activate', extension)
     }
 
+    /**
+     * @description 每个插件的入口文件extension.js必须导出一个instance对象实现extensionInstance接口
+     * @param extension
+     */
     async doActivateExtension(extension: IMainExtension) {
-        let { activate, beforeClose, actualEntrance } = require(extension.storage)
         try {
-            await activate()
-            ExtensionManager.registerCloseFunction(beforeClose)
+            let { instance } = await import(extension.storage)
+            await instance.activate()
+            let worker = undefined
+            if (instance.workerEntrance) {
+                worker = new Worker(instance.workerEntrance)
+            }
             ExtensionActivator.extensionInstanceManagers.set(extension.identifier.id, {
                 identifier: extension.identifier,
-                worker: new Worker(actualEntrance),
+                worker: worker,
                 instance: {
-                    activate: activate,
-                    beforeClose: beforeClose,
-                    actualEntrance: actualEntrance,
+                    activate: instance.activate,
+                    beforeClose: instance.beforeClose,
+                    workerEntrance: instance.workerEntrance,
                 },
             })
         } catch (e: any) {
@@ -52,7 +65,7 @@ export class ExtensionActivator {
         try {
             if (instance) {
                 instance.instance.beforeClose()
-                instance.worker.terminate()
+                if (instance.worker) instance.worker.terminate()
             }
         } catch (e: any) {
             throw e
@@ -61,7 +74,7 @@ export class ExtensionActivator {
 
     static beforeClose() {
         ExtensionActivator.extensionInstanceManagers.forEach(
-            (instance: extensionInstanceManager, extensionId: string) => {
+            (instance: IExtensionInstanceManager, extensionId: string) => {
                 ExtensionActivator.terminateExtensionInstance(extensionId)
             }
         )
